Update the product variant precentage type

+++ Presentation/dto/ProductVariantDto.cs
@@ -7,7 +7,7 @@ namespace api.Presentation.dto
         public Guid Id { get; set; }
         public Guid ProductId { get; set; }
         public string? Name { get; set; }
-        public decimal Percentage { get; set; }
+        public int Percentage { get; set; }
         public Guid VariantId { get; set; }
     }
     
@@ -15,7 +15,7 @@ namespace api.Presentation.dto
     {
         [StringLength(maximumLength:50,MinimumLength =1 ,ErrorMessage= "name must not be empty")]
         public string Name { get; set; }
-        public ushort Percentage { get; set; } = 1;
+        public int Percentage { get; set; } = 1;
         public Guid VariantId { get; set; }
     }
     
@@ -23,7 +23,7 @@ namespace api.Presentation.dto
     {
         [StringLength(maximumLength:50 ,ErrorMessage= "name must not  be empty")]
         public string? Name { get; set; }
-        public decimal Percentage { get; set; }
+        public int Percentage { get; set; }
         public string?  VariantName { get; set; } 
     }
 }
